+++ src/main/java/io/github/qudtlib/maven/rdfio/pipeline/step/ShaclInferStep.java
@@ -9,6 +9,7 @@ import io.github.qudtlib.maven.rdfio.pipeline.step.support.Inferred;
 import io.github.qudtlib.maven.rdfio.pipeline.step.support.InputsComponent;
 import io.github.qudtlib.maven.rdfio.pipeline.step.support.ParsingHelper;
 import io.github.qudtlib.maven.rdfio.pipeline.support.ConfigurationParseException;
+import io.github.qudtlib.maven.rdfio.pipeline.support.PipelineConfigurationExeception;
 import io.github.qudtlib.maven.rdfio.pipeline.support.VariableResolver;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
@@ -314,7 +315,16 @@ public class ShaclInferStep implements Step {
             allGraphs.addAll(additionalGraphs);
             allGraphs.addAll(inputsComponent.getAllInputGraphs(dataset, state));
             if (allGraphs != null) {
-                allGraphs.forEach(g -> dataModel.add(dataset.getNamedModel(g)));
+                allGraphs.forEach(
+                        g -> {
+                            if (!dataset.containsNamedModel(g)
+                                    && !state.getShaclFunctionsGraph().equals(g)) {
+                                throw new PipelineConfigurationExeception(
+                                        "No graph %s found in dataset, cannot use in shaclInfer"
+                                                .formatted(g));
+                            }
+                            dataModel.add(dataset.getNamedModel(g));
+                        });
             }
             entries.addAll(PipelineHelper.formatGraphs(allGraphs));
         }